requirements.txt
python-dotenv
requests
translate-api